gaas-bot/pyproject.toml
[project]
name = "gaas-bot"
version = "0.1.0"
description = "Maintenance automation for GaaS — issue resolution, audits, and more"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "claude-agent-sdk>=0.1.45",
    "githubkit[auth-app]>=0.12",
    "jinja2>=3.1.6",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
gaas-bot = "gaas_bot.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gaas_bot"]
